More tests changes
