+++ Cargo.toml
@@ -80,6 +80,40 @@ strip = "symbols"
 [package.metadata.cargo-udeps.ignore]
 normal = [ "async-trait", "derive_builder" ]
 
+# Block examples
+
+[[example]]
+name = "block_no_payload"
+path = "examples/block/00_block_no_payload.rs"
+
+[[example]]
+name = "block_confirmation_time"
+path = "examples/block/01_block_confirmation_time.rs"
+
+[[example]]
+name = "block_custom_parents"
+path = "examples/block/02_block_custom_parents.rs"
+
+[[example]]
+name = "custom_inputs"
+path = "examples/block/custom_inputs.rs"
+
+[[example]]
+name = "custom_payload"
+path = "examples/block/custom_payload.rs"
+
+[[example]]
+name = "data_block"
+path = "examples/block/data_block.rs"
+
+[[example]]
+name = "output"
+path = "examples/block/output.rs"
+
+[[example]]
+name = "transaction"
+path = "examples/block/transaction.rs"
+
 # Node API core examples
 
 [[example]]
@@ -229,38 +263,6 @@ required-features = [ "ledger_nano" ]
 [[example]]
 name = "stronghold"
 required-features = [ "stronghold" ]
-
-[[example]]
-name = "block_time"
-path = "examples/block/block_time.rs"
-
-[[example]]
-name = "custom_inputs"
-path = "examples/block/custom_inputs.rs"
-
-[[example]]
-name = "custom_parent"
-path = "examples/block/custom_parent.rs"
-
-[[example]]
-name = "custom_payload"
-path = "examples/block/custom_payload.rs"
-
-[[example]]
-name = "data_block"
-path = "examples/block/data_block.rs"
-
-[[example]]
-name = "output"
-path = "examples/block/output.rs"
-
-[[example]]
-name = "simple_block"
-path = "examples/block/simple_block.rs"
-
-[[example]]
-name = "transaction"
-path = "examples/block/transaction.rs"
 
 [[example]]
 name = "0_address_generation"
