+++ examples/block/02_block_custom_parents.rs
@@ -1,0 +1,34 @@
+// Copyright 2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+//! This example sends a block, with custom parents, which can be used for promoting.
+//! Run: `cargo run --example block_custom_parents --release -- [NODE URL]`.
+
+use iota_client::{Client, Result};
+
+#[tokio::main]
+async fn main() -> Result<()> {
+    // Take the node URL from command line argument or use one from env as default.
+    let node_url = std::env::args().nth(1).unwrap_or_else(|| {
+        // This example uses dotenv, which is not safe for use in production.
+        dotenv::dotenv().ok();
+        std::env::var("NODE_URL").unwrap()
+    });
+
+    // Create a client with that node.
+    let client = Client::builder().with_node(&node_url)?.finish()?;
+
+    // Use tips as custom parents.
+    let parents = client.get_tips().await?;
+
+    // Create the block with custom parents.
+    let block = client.block().with_parents(parents)?.finish().await?;
+
+    println!(
+        "Block sent: {}/block/{}",
+        std::env::var("EXPLORER_URL").unwrap(),
+        block.id()
+    );
+
+    Ok(())
+}
