--- examples/block/custom_parent.rs
@@ -1,32 +1,0 @@
-// Copyright 2021 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-//! cargo run --example custom_parent --release
-
-use iota_client::{Client, Result};
-
-/// In this example we will define a custom block parent which be used for promoting
-
-#[tokio::main]
-async fn main() -> Result<()> {
-    dotenv::dotenv().ok();
-
-    let node_url = std::env::var("NODE_URL").unwrap();
-
-    // Create a client instance
-    let client = Client::builder()
-        .with_node(&node_url)? // Insert your node URL here
-        .finish()?;
-
-    let custom_parent = client.get_tips().await?[0];
-
-    let block = client.block().with_parents(vec![custom_parent])?.finish().await?;
-
-    println!(
-        "Empty block sent: {}/block/{}",
-        std::env::var("EXPLORER_URL").unwrap(),
-        block.id()
-    );
-
-    Ok(())
-}
